[Refactor] qa module to opensearch
Refactors qa module from legacy namespace to opensearch.
