+++ qa/kerberos/src/itest/java/org/opensearch/hadoop/qa/kerberos/AbstractClusterVerificationTests.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.elasticsearch.hadoop.qa.kerberos;
+package org.opensearch.hadoop.qa.kerberos;
 
 import java.io.BufferedReader;
 import java.io.File;
