+++ qa/kerberos/src/main/resources/hive/load_to_opensearch.sql
