+++ qa/kerberos/src/main/resources/hive/read_from_opensearch.sql
