+++ qa/kerberos/src/main/resources/pig/load_to_opensearch.pig
