+++ qa/kerberos/src/main/resources/pig/read_from_opensearch.pig
