+++ qa/kerberos/src/main/scala/org/elasticsearch/hadoop/qa/kerberos/spark/LoadToES.scala
@@ -23,9 +23,9 @@ import java.security.PrivilegedExceptionAction
 import org.apache.spark.SparkConf
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.types.{StringType, StructField, StructType}
-import org.elasticsearch.hadoop.qa.kerberos.security.KeytabLogin
 import org.elasticsearch.spark._
 import org.elasticsearch.spark.sql._
+import org.opensearch.hadoop.qa.kerberos.security.KeytabLogin
 
 class LoadToES(args: Array[String]) {
